relocated old scripts
