+++ Assets/Editor/ReadOnlyDrawer.cs
@@ -1,4 +1,4 @@
-using DungeonGenerator.Extensions;
+using DungeonGenerator.Old.Extensions;
 using UnityEngine;
 using UnityEditor;
