+++ Assets/Scripts/Managers.meta
@@ -1,8 +1,3 @@
-fileFormatVersion: 2
-guid: 7794cd0380ac08d4a8742e25dfd8e4e8
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant: 
+﻿fileFormatVersion: 2
+guid: 17c2a324b2a847d889d7eccb20614163
+timeCreated: 1651458894
