+++ Assets/Scripts/Old/Data/DungeonData.cs
@@ -1,7 +1,7 @@
-﻿using DungeonGenerator.Extensions;
+﻿using DungeonGenerator.Old.Extensions;
 using UnityEngine;
 
-namespace DungeonGenerator.Data
+namespace DungeonGenerator.Old.Data
 {
     /// <summary>
     /// Contains the amount of rooms for the dungeon.
