+++ Assets/Scripts/Old/Managers.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 7794cd0380ac08d4a8742e25dfd8e4e8
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
