+++ Assets/Scripts/Old/Managers/DungeonManagerOld.cs
@@ -1,11 +1,11 @@
 using System.Collections;
 using System.Collections.Generic;
-using DungeonGenerator.Data;
-using DungeonGenerator.Objects;
-using DungeonGenerator.Types;
+using DungeonGenerator.Old.Data;
+using DungeonGenerator.Old.Objects;
+using DungeonGenerator.Old.Types;
 using UnityEngine;
 
-namespace DungeonGenerator.Managers
+namespace DungeonGenerator.Old.Managers
 {
     public class DungeonManagerOld : MonoBehaviour
     {
