+++ Assets/Scripts/Old/Types/Orientation.cs
@@ -1,4 +1,4 @@
-﻿namespace DungeonGenerator.Types
+﻿namespace DungeonGenerator.Old.Types
 {
     public enum Orientation
     {
