+++ Assets/Scripts/Old/Types/RoomType.cs
@@ -1,4 +1,4 @@
-﻿namespace DungeonGenerator.Types
+﻿namespace DungeonGenerator.Old.Types
 {
     public enum RoomType
     {
